requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==0.5.2
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
black==21.4b2
blessed==1.20.0
braceexpand==0.1.7
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
-e git+ssh://git@github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
cloudpickle==2.2.1
contourpy==1.1.1
controlnet-aux==0.0.3
cycler==0.11.0
Cython==3.0.2
detectron2 @ file:///root/.packages/detectron2-0.6%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=06e62c02f6b4ec7ab49608d5cc85fee2253825db44e8d9bf6dec870e9f255a26
diffusers==0.18.0
docker-pycreds==0.4.0
easydict==1.10
einops==0.6.1
exceptiongroup==1.1.3
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.12.4
fire==0.5.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flaskwebgui==0.3.5
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.1
google-auth-oauthlib==1.0.0
gpustat==1.1.1
gradio==3.45.0
gradio_client==0.5.2
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
hydra-core==1.3.2
idna==3.4
imageio==2.31.4
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kornia==0.5.0
lama-cleaner==1.2.3
lazy_loader==0.3
lightning-utilities==0.9.0
loguru==0.7.2
lpips==0.1.4
madgrad==1.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1.1
numpy==1.23.5
nvidia-ml-py==12.535.108
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.3.0
open-clip-torch==1.3.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
orjson==3.9.7
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pathtools==0.1.2
piexif==1.1.3
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
portalocker==2.8.2
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.4.1
pydantic_core==2.10.1
pydot==1.4.2
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.7.1
python-multipart==0.0.6
python-socketio==5.9.0
pytorch-lightning==1.9.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
# Editable install with no version control (segment-anything==1.0)
-e /root/.packages/segment-anything-main
semantic-version==2.10.0
sentry-sdk==1.31.0
setproctitle==1.3.2
shapely==2.0.1
simple-websocket==0.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.7
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.10.1
torchaudio==0.10.1
torchmetrics==1.2.0
torchvision==0.11.2
tqdm==4.66.1
transformers==4.27.4
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
wandb==0.15.11
wcwidth==0.2.6
webdataset==0.2.57
websockets==11.0.3
Werkzeug==2.3.7
whichcraft==0.6.1
wldhx.yadisk-direct==0.0.6
wsproto==1.2.0
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
